Some simplifications
`perform` now mutates the object, but is bound late so that it
may still be used on a clone.

Ids are removed from event output; it seems unreasonable to expect
clients to track these ids when we already pretty much guarantee
that two entities of the same type never overlap.