Incrementally update node info
Update __nodes in O(1) time. This kind of code brings with it a
nontrivial amount of cognitive overhead, so in debug mode, it is
also verified against a freshly computed table.

(what a shame debug mode is the *default* in Python...)

The performance impact is not yet visible because the largest
troublemaker is still the event system.