Give each rule its own MoveCache
Doing this allows moves and kinds can now be represented by
simpler objects in the IncrementalMoveCache, as they are now
inherently namespaced by rule.

It also somewhat legitimizes the use of inheritance for Rules;
though it was never my intent initially for Rules to be stateful,
they technically *have* been stateful ever since the incrementalization
update. (in the sense that calling one member function with a fixed
set of arguments may have different results based on the history
of previously called methods)