Further simply Rule implementations
A boilerplatey pattern has arisen in my Rule implementations
suggesting that there is yet a simpler model for them.  This merely
adds an intermediate class which factors out the common details.
(gasp, a class heirarchy?!)

Somewhat annoyingly, `intialize_moves` was part of that boilerplate.
"Annoyingly" because one of the purposes of that method was for
validating the others!  Now that they all share a large bulk of their
implementation, it is unclear what sort of errors will actually
be able to be caught by that `--validate-every` flag I just added...