Baby step: Ensure code is "layers-aware"
Audited all existing code to make sure it is aware that not
all vacancies are necessarily divacancies.

This also makes rules.py stop inspecting the stuff in __nodes
and instead call node-testing functions helper functions in State,
which is a bit of a change from the plans I had earlier.

---

Those plans were to get rid of `State.{pop,new}_{trefoil,vacancy}` and
implement all that logic in Rules, on the basis that Rules and State
are inseparably coupled; in the majority of cases, adding a new rule
requires adding stuff to State, possibly impacting other rules.

The plans changed because the ergonomics were terrible (and bad
ergonomics + loose type system = high potential for mistakes)!
I think I can accept the fact that State and Rules are highly
coupled so long as it enables me to write stuff like
`filter(state.is_divacancy, nodes)`.