Use entity lists for initialization
This not only improves the efficiency of initialization, but it
also improves the effectiveness of the --validate-every flag by
once again making different code paths produce the "expected"
and "actual" move sets.

Lo and behold, it helped catch a bug in my recently added
monovacancy rules! (deleting the last monovacancy in a node
was leaving behind a "vacancy" with 0 layers)